+++ src/main/java/by/gurinovich/bamper/security/expressions/CustomSecurityExpression.java
@@ -13,7 +13,7 @@ public class CustomSecurityExpression {
 
     private final UserService userService;
 
-    public boolean isPostOwner(final Integer post_id){
+    public boolean isPostOwner(final Long post_id){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         JWTEntity jwtEntity = (JWTEntity) authentication.getPrincipal();
         System.out.println(jwtEntity.getUsername());
